rename module, add error messages
